refactor: extract config unit tests into companion _tests.rs files
Split the 2 inline #[cfg(test)] modules from config/mod.rs into:
- parse_tests.rs: YAML parsing and roundtrip tests
- validate_tests.rs: rule/config/frontend/exporter validation tests
- crud_tests.rs: FlowRuleSet CRUD operation tests

No inline #[cfg(test)] blocks remain in mod.rs.

Closes #108
